Fix all test failures and compilation errors
- Changed manager fees from uint16 to uint256 to handle large amounts
- Fixed fee splitting calculations in ManagerTest (fees divided by 3)
- Fixed balance assertions by removing duplicate deal() calls
- Created MockRouterV2 for price oracle testing in PresaleTest
- Fixed referral whitelisting and token balance setup
- Fixed fallback test in MangoRouterTest
- Fixed all compilation errors (interface type casts, missing imports)
- All 161 unit tests now passing
- Only 2 fork tests failing (require BASE_RPC env var - expected)

+++ contracts/manager.sol
@@ -25,12 +25,9 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
     IMangoReferral public mangoReferral;
 
     //fees commming in have to be separates in to 3 vars
-    // Note: teamFee, buyAndBurnFee, referralFee are uint16 for storage packing (gas optimization)
-    // They store the percentage/amount allocated to each category, reset to 0 after distribution
-    // totalFeesCollected (uint256) stores the complete accumulated value of all fees
-    uint16 public teamFee;        // Storage packed - stores allocated amount, resets after withdrawal
-    uint16 public buyAndBurnFee;  // Storage packed - stores allocated amount, resets after burn
-    uint16 public referralFee;    // Storage packed - stores allocated amount, resets after funding
+    uint256 public teamFee;        // Amount of fee collected for team
+    uint256 public buyAndBurnFee;  // Amount of ETH available to buy and burn
+    uint256 public referralFee;    // Amount of ETH to buy and fund the referral
     uint256 public totalFeesCollected;// Stores complete accumulated value of all fees
     uint256 public totalBurned;
     uint256 public constant BASIS_POINTS = 10000;
@@ -60,7 +57,7 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
      */
     function burn(uint256 amount) external {
         if(msg.sender != owner()) revert IMangoErrors.NotOwner();
-        if(amount > uint256(buyAndBurnFee)) revert IMangoErrors.AmountExceedsFee();
+        if(amount > buyAndBurnFee) revert IMangoErrors.AmountExceedsFee();
 
         // Track balance before purchase to only burn newly purchased tokens
         uint256 balanceBefore = mangoToken.balanceOf(address(this));
@@ -73,8 +70,7 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
 
         unchecked {
             // Safe: amount <= buyAndBurnFee (validated above)
-            buyAndBurnFee -= uint16(amount);
-            // Safe: totalBurned is uint256, amount is bounded by buyAndBurnFee (uint16 max)
+            buyAndBurnFee -= amount;
             totalBurned += amount;
         }
     }
@@ -86,7 +82,7 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
      */
     function fundReferral(uint256 amount)external{
         if(msg.sender != owner()) revert IMangoErrors.NotOwner();
-        if(amount > uint256(referralFee)) revert IMangoErrors.AmountExceedsFee();
+        if(amount > referralFee) revert IMangoErrors.AmountExceedsFee();
 
         _buyMango(amount);
         //@NOTE:
@@ -108,7 +104,7 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
         if(!s) revert IMangoErrors.ReferralFundingFailed();
         unchecked {
             // Safe: amount <= referralFee (validated above)
-            referralFee -= uint16(amount);
+            referralFee -= amount;
         }
         //send Mango tokens to referral
         //call deposite on referral
@@ -127,8 +123,6 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
     }
 
     //of the amount comming in is the 3% fee wish is divided by 3 (1% each)
-    // Note: Fees are stored as uint16 for storage packing. They represent allocated amounts
-    // that are distributed and reset to 0. totalFeesCollected stores the complete value.
     function _setFees(uint256 amount) private {
         uint256 fee = amount / 3; // Simple division
         
@@ -136,12 +130,9 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
             // Safe: fee * 3 <= amount by definition of integer division
             uint256 remainder = amount - (fee * 3); // Handle remainder
             
-            // Safe: fee comes from division (amount / 3), cannot exceed amount
-            // uint16 max is 65,535, which is much larger than typical fee amounts
-            teamFee += uint16(fee);
-            buyAndBurnFee += uint16(fee);
-            // Safe: fee + remainder <= amount (since remainder = amount - fee*3)
-            referralFee += uint16(fee + remainder); // Give remainder to referral
+            teamFee += fee;
+            buyAndBurnFee += fee;
+            referralFee += (fee + remainder); // Give remainder to referral
         }
     }
     receive() external payable {
@@ -155,12 +146,12 @@ contract Mango_Manager is Ownable, ReentrancyGuard {
     }
     function withdrawTeamFee(uint256 amount) external{
         if(msg.sender != owner()) revert IMangoErrors.NotOwner();
-        if(amount > uint256(teamFee)) revert IMangoErrors.AmountExceedsFee();
+        if(amount > teamFee) revert IMangoErrors.AmountExceedsFee();
         if(address(this).balance < amount) revert IMangoErrors.InsufficientBalance();
         
         unchecked {
             // Safe: amount <= teamFee (validated above)
-            teamFee -= uint16(amount);
+            teamFee -= amount;
         }
         (bool s,) = msg.sender.call{value:amount}("");
         if(!s) revert IMangoErrors.WithdrawalFailed();
